Create README.md
Design an air-suction manipulator with a negative pressure generator, a relief valve and an adsorption sucker as the core to realize a grasping device with the characteristics of an unmanned helicopter. This grasping device communicates with the remote vision system and cooperates with the binocular camera on the unmanned helicopter for image recognition, converting the image coordinates of the target object into world coordinates, so that the device can complete corresponding tasks such as accurate search, pickup and delivery.